requirements_render.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
gunicorn==21.2.0
python-docx==1.0.1
beautifulsoup4==4.12.2
playwright==1.42.0
openpyxl==3.1.2
requests==2.31.0
lxml-stubs==0.4.0
html5lib==1.1
markdownify==0.11.6
Pillow==10.1.0
